Restructure package
