+++ dracobot2/utils/__init__.py
@@ -1,0 +1,2 @@
+from .resources import *
+from .msg_mappings import *
