+++ dracobot2/utils/msg_mappings.py
@@ -1,37 +1,13 @@
 import telegram
 from telegram import InputMediaPhoto, InputMediaAudio, InputMediaDocument, InputMediaVideo
 from telegram.ext import Filters
-from models import MessageMapping, MsgFrom
 from sqlalchemy import or_, and_
-from resources import *
+from dracobot2.models import MessageMapping, MsgFrom
+from dracobot2.resources import *
+from .resources import *
 
 
 SUPPORTED_MESSAGE_FILTERS = Filters.photo | Filters.sticker | Filters.document | Filters.video | Filters.video_note | Filters.audio | Filters.voice | Filters.text
-
-
-def format_message(message_text, message_from=MsgFrom.DRAGON, is_prefix=False, is_edited=False):
-    receiver_type = 'Dragon' if message_from == MsgFrom.DRAGON else 'Trainer' if message_from == MsgFrom.TRAINER else 'Admin'
-
-    if is_edited:
-        message_text += ' (edited)'
-
-    if is_prefix:
-        ret_msg = "%s:\n%s" % (receiver_type, message_text)
-    else:
-        if message_text is not None and len(message_text) > 0:
-            ret_msg = "%s\n- %s" % (message_text, receiver_type)
-        else:
-            ret_msg = "- %s" % (receiver_type)
-    return ret_msg
-
-
-def format_registered_message(user):
-    if user is None:
-        return "Not assigned " + CROSS
-    elif user.registered:
-        return "Registered " + GREEN_STATUS
-    else:
-        return "Not registered " + RED_STATUS
 
 
 def get_highest_resolution(photos):
