+++ setup.py
@@ -1,5 +1,5 @@
-from config import SQL_ENGINE
-from models import *
+from dracobot2.config import SQL_ENGINE
+from dracobot2.models import *
 
 print("Creating Database")
 Base.metadata.create_all(SQL_ENGINE)
